Enhances view structure and script loading
Moves script and style section definitions to the top of the view files.

This improves the organization and maintainability of the views by centralizing the inclusion of styles and scripts. It also ensures scripts are defined before being used, preventing potential errors.

Also, convert inline script in Towns Index view to external js file and include interaction enhancements to the town admin table.